+++ client/src/pages/Dashboard/Products/AddSizes/AddSize.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function AddSize() {
+  return <div>AddSize</div>;
+}
